services/order/pyproject.toml

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ecom-order"
version = "0.1.0"
description = "Order service (checkout + inventory reservation + Kafka)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.5.0",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.0",
    "psycopg[binary]>=3.1.10",
    "redis>=5.0.4",
    "httpx>=0.27.0",
    "kafka-python>=2.0.2",
    "PyJWT>=2.9.0",
]

[tool.setuptools.packages.find]
where = [""]
include = ["app*"]
